+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Unreleased
 - optionally, index only residues with 'good quality' (e.g., based on B-factor values)
 - make `StructureWriterImpl` configurable
 - support indexing of non-archived structures by reading from local file or URL (this allows to index e.g. AlphaFold data and search therein)
-- filter by TargetList (`PDB`, `MODELS`, or `ALL`) to restrict overall search space of queries
+- filter by SearchSpace (`PDB`, `MODELS`, or `ALL`) to restrict overall search space of queries
 
 strucmotif-search 0.12.8
 -------------